+++ perception/object_merger/include/autoware/object_merger/data_association/solver/gnn_solver.hpp
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
-#define OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
+#ifndef AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
+#define AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
 
-#include "object_merger/data_association/solver/gnn_solver_interface.hpp"
-#include "object_merger/data_association/solver/mu_successive_shortest_path.hpp"
-#include "object_merger/data_association/solver/successive_shortest_path.hpp"
+#include "autoware/object_merger/data_association/solver/gnn_solver_interface.hpp"
+#include "autoware/object_merger/data_association/solver/mu_successive_shortest_path.hpp"
 
-#endif  // OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
+#endif  // AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_HPP_
