+++ perception/object_merger/include/autoware/object_merger/data_association/solver/gnn_solver_interface.hpp
@@ -12,26 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__SUCCESSIVE_SHORTEST_PATH_HPP_
-#define OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__SUCCESSIVE_SHORTEST_PATH_HPP_
-
-#include "object_merger/data_association/solver/gnn_solver_interface.hpp"
+#ifndef AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
+#define AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
 
 #include <unordered_map>
 #include <vector>
 
 namespace gnn_solver
 {
-class SSP : public GnnSolverInterface
+class GnnSolverInterface
 {
 public:
-  SSP() = default;
-  ~SSP() = default;
+  GnnSolverInterface() = default;
+  virtual ~GnnSolverInterface() = default;
 
-  void maximizeLinearAssignment(
+  virtual void maximizeLinearAssignment(
     const std::vector<std::vector<double>> & cost, std::unordered_map<int, int> * direct_assignment,
-    std::unordered_map<int, int> * reverse_assignment) override;
+    std::unordered_map<int, int> * reverse_assignment) = 0;
 };
 }  // namespace gnn_solver
 
-#endif  // OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__SUCCESSIVE_SHORTEST_PATH_HPP_
+#endif  // AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
