+++ perception/object_merger/include/autoware/object_merger/data_association/solver/mu_successive_shortest_path.hpp
@@ -12,30 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "object_merger/data_association/solver/mu_successive_shortest_path.hpp"
+#ifndef AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__MU_SUCCESSIVE_SHORTEST_PATH_HPP_
+#define AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__MU_SUCCESSIVE_SHORTEST_PATH_HPP_
 
-#include <mussp/mussp.h>
+#include "autoware/object_merger/data_association/solver/gnn_solver_interface.hpp"
 
-#include <array>
-#include <cassert>
-#include <fstream>
-#include <iostream>
-#include <string>
 #include <unordered_map>
 #include <vector>
 
 namespace gnn_solver
 {
-void MuSSP::maximizeLinearAssignment(
-  const std::vector<std::vector<double>> & cost, std::unordered_map<int, int> * direct_assignment,
-  std::unordered_map<int, int> * reverse_assignment)
+class MuSSP : public GnnSolverInterface
 {
-  // Terminate if the graph is empty
-  if (cost.size() == 0 || cost.at(0).size() == 0) {
-    return;
-  }
+public:
+  MuSSP() = default;
+  ~MuSSP() = default;
 
-  // Solve DA by muSSP
-  solve_muSSP(cost, direct_assignment, reverse_assignment);
-}
+  void maximizeLinearAssignment(
+    const std::vector<std::vector<double>> & cost, std::unordered_map<int, int> * direct_assignment,
+    std::unordered_map<int, int> * reverse_assignment) override;
+};
 }  // namespace gnn_solver
+
+#endif  // AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__MU_SUCCESSIVE_SHORTEST_PATH_HPP_
