+++ perception/object_merger/include/autoware/object_merger/data_association/solver/successive_shortest_path.hpp
@@ -12,24 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
-#define OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
+#ifndef AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__SUCCESSIVE_SHORTEST_PATH_HPP_
+#define AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__SUCCESSIVE_SHORTEST_PATH_HPP_
+
+#include "autoware/object_merger/data_association/solver/gnn_solver_interface.hpp"
 
 #include <unordered_map>
 #include <vector>
 
 namespace gnn_solver
 {
-class GnnSolverInterface
+class SSP : public GnnSolverInterface
 {
 public:
-  GnnSolverInterface() = default;
-  virtual ~GnnSolverInterface() = default;
+  SSP() = default;
+  ~SSP() = default;
 
-  virtual void maximizeLinearAssignment(
+  void maximizeLinearAssignment(
     const std::vector<std::vector<double>> & cost, std::unordered_map<int, int> * direct_assignment,
-    std::unordered_map<int, int> * reverse_assignment) = 0;
+    std::unordered_map<int, int> * reverse_assignment) override;
 };
 }  // namespace gnn_solver
 
-#endif  // OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__GNN_SOLVER_INTERFACE_HPP_
+#endif  // AUTOWARE__OBJECT_MERGER__DATA_ASSOCIATION__SOLVER__SUCCESSIVE_SHORTEST_PATH_HPP_
