+++ perception/object_merger/src/data_association/data_association.cpp
@@ -12,11 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "object_merger/data_association/data_association.hpp"
+#include "autoware/object_merger/data_association/data_association.hpp"
 
+#include "autoware/object_merger/data_association/solver/gnn_solver.hpp"
 #include "autoware/universe_utils/geometry/geometry.hpp"
-#include "object_merger/data_association/solver/gnn_solver.hpp"
-#include "object_merger/utils/utils.hpp"
 #include "object_recognition_utils/object_recognition_utils.hpp"
 
 #include <algorithm>
