+++ perception/object_merger/src/data_association/successive_shortest_path/successive_shortest_path.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "object_merger/data_association/solver/successive_shortest_path.hpp"
+#include "autoware/object_merger/data_association/solver/successive_shortest_path.hpp"
 
 #include <algorithm>
 #include <cassert>
